requirements.txt
paddlepaddle-gpu==2.4.2
face_alignment==1.3.5
numpy==1.23.4
imageio==2.19.3
imageio-ffmpeg==0.4.7
librosa==0.9.2
numba
resampy==0.3.1
pydub==0.25.1
scipy==1.5.3
kornia==0.6.8
tqdm
yacs==0.1.8
pyyaml
joblib==1.1.0
scikit-image==0.19.3
basicsr==1.4.2
facexlib==0.2.5
gradio
gfpgan
asgiref==3.4.1  # Latest version known to be compatible with Django and Python 3.8
Django==3.2.10  # Latest LTS version compatible with Python 3.8
django-filter==2.4.0
djangorestframework==3.12.4
Markdown==3.3.4
dlib-bin
celery[redis]    